# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedPrep AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UserProfile {
  id: string;
  email: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: string;
  /** @nullable */
  collegeName?: string | null;
  /** @nullable */
  yearOfStudy?: number | null;
  /** @nullable */
  examTarget?: string | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  fullName?: string;
  collegeName?: string;
  yearOfStudy?: number;
  examTarget?: string;
}

export type JobType = typeof JobType[keyof typeof JobType];


export const JobType = {
  PDF_ANALYSIS: 'PDF_ANALYSIS',
  NOTES_GENERATION: 'NOTES_GENERATION',
  SUMMARY_GENERATION: 'SUMMARY_GENERATION',
  FLASHCARD_GENERATION: 'FLASHCARD_GENERATION',
  QUIZ_GENERATION: 'QUIZ_GENERATION',
  REVISION_PDF: 'REVISION_PDF',
} as const;

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  PENDING: 'PENDING',
  RUNNING: 'RUNNING',
  SUCCEEDED: 'SUCCEEDED',
  FAILED: 'FAILED',
} as const;

export interface ProcessingJob {
  id: string;
  /** @nullable */
  pdfId?: string | null;
  userId: string;
  type: JobType;
  status: JobStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalPdfs: number;
  processingPdfs: number;
  completedPdfs: number;
  failedPdfs?: number;
  totalQuestions: number;
  recentJobs: ProcessingJob[];
}

export type PdfStatus = typeof PdfStatus[keyof typeof PdfStatus];


export const PdfStatus = {
  UPLOADED: 'UPLOADED',
  QUEUED: 'QUEUED',
  PROCESSING: 'PROCESSING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
} as const;

export interface Pdf {
  id: string;
  userId: string;
  /** @nullable */
  subjectId?: string | null;
  fileName: string;
  storagePath: string;
  fileSizeBytes: number;
  /** @nullable */
  pageCount?: number | null;
  status: PdfStatus;
  /** @nullable */
  errorMessage?: string | null;
  uploadedAt: string;
  /** @nullable */
  processedAt?: string | null;
}

export interface PdfList {
  items: Pdf[];
  total: number;
  page: number;
  limit: number;
}

export interface PdfStatusResponse {
  id: string;
  status: PdfStatus;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  processedAt?: string | null;
}

export interface PdfRegistration {
  fileName: string;
  storagePath: string;
  fileSizeBytes: number;
  subjectId?: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type ListPdfsParams = {
status?: ListPdfsStatus;
/**
 * @nullable
 */
subjectId?: string | null;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};

export type ListPdfsStatus = typeof ListPdfsStatus[keyof typeof ListPdfsStatus];


export const ListPdfsStatus = {
  UPLOADED: 'UPLOADED',
  QUEUED: 'QUEUED',
  PROCESSING: 'PROCESSING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
} as const;

export type ListJobsParams = {
status?: ListJobsStatus;
/**
 * @nullable
 */
pdfId?: string | null;
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};

export type ListJobsStatus = typeof ListJobsStatus[keyof typeof ListJobsStatus];


export const ListJobsStatus = {
  PENDING: 'PENDING',
  RUNNING: 'RUNNING',
  SUCCEEDED: 'SUCCEEDED',
  FAILED: 'FAILED',
} as const;