# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedPrep AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProcessingJob } from './processingJob';

export interface DashboardStats {
  totalPdfs: number;
  processingPdfs: number;
  completedPdfs: number;
  failedPdfs?: number;
  totalQuestions: number;
  recentJobs: ProcessingJob[];
}