# lib/api-zod/src/generated/types/jobStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedPrep AI API specification
 * OpenAPI spec version: 0.1.0
 */

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  PENDING: 'PENDING',
  RUNNING: 'RUNNING',
  SUCCEEDED: 'SUCCEEDED',
  FAILED: 'FAILED',
} as const;