# lib/api-zod/src/generated/types/jobType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedPrep AI API specification
 * OpenAPI spec version: 0.1.0
 */

export type JobType = typeof JobType[keyof typeof JobType];


export const JobType = {
  PDF_ANALYSIS: 'PDF_ANALYSIS',
  NOTES_GENERATION: 'NOTES_GENERATION',
  SUMMARY_GENERATION: 'SUMMARY_GENERATION',
  FLASHCARD_GENERATION: 'FLASHCARD_GENERATION',
  QUIZ_GENERATION: 'QUIZ_GENERATION',
  REVISION_PDF: 'REVISION_PDF',
} as const;