# lib/api-zod/src/generated/types/pdf.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedPrep AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PdfStatus } from './pdfStatus';

export interface Pdf {
  id: string;
  userId: string;
  /** @nullable */
  subjectId?: string | null;
  fileName: string;
  storagePath: string;
  fileSizeBytes: number;
  /** @nullable */
  pageCount?: number | null;
  status: PdfStatus;
  /** @nullable */
  errorMessage?: string | null;
  uploadedAt: Date;
  /** @nullable */
  processedAt?: Date | null;
}