# lib/api-zod/src/generated/types/pdfList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedPrep AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Pdf } from './pdf';

export interface PdfList {
  items: Pdf[];
  total: number;
  page: number;
  limit: number;
}