# lib/api-zod/src/generated/types/pdfStatusResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedPrep AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PdfStatus } from './pdfStatus';

export interface PdfStatusResponse {
  id: string;
  status: PdfStatus;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  processedAt?: Date | null;
}