# lib/api-zod/src/generated/types/processingJob.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedPrep AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from './jobStatus';
import type { JobType } from './jobType';

export interface ProcessingJob {
  id: string;
  /** @nullable */
  pdfId?: string | null;
  userId: string;
  type: JobType;
  status: JobStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  startedAt?: Date | null;
  /** @nullable */
  completedAt?: Date | null;
  createdAt: Date;
}