# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedPrep AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: string;
  email: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: string;
  /** @nullable */
  collegeName?: string | null;
  /** @nullable */
  yearOfStudy?: number | null;
  /** @nullable */
  examTarget?: string | null;
  createdAt: Date;
}